// server/enums/planActionStatus.ts
enum PlanActionStatus {
  PLAN_OVERDUE = 'PLAN_OVERDUE',
  PLAN_DUE = 'PLAN_DUE',
  REVIEW_OVERDUE = 'REVIEW_OVERDUE',
  REVIEW_DUE = 'REVIEW_DUE',
  ACTIVE_PLAN = 'ACTIVE_PLAN',
  NEEDS_PLAN = 'NEEDS_PLAN',
  INACTIVE_PLAN = 'INACTIVE_PLAN',
  PLAN_DECLINED = 'PLAN_DECLINED',
  NO_PLAN = 'NO_PLAN',
}

export default PlanActionStatus